+++ src/evilstructs/__init__.py
@@ -1,0 +1,5 @@
+"""All the data structures I don't want to write twice."""
+
+from .frozendict import frozendict
+
+__all__ = ["frozendict"]