+++ CHANGELOG.md
@@ -8,5 +8,7 @@ Breaking changes:
   should still be preserved.
 
 Major changes:
+
 - Added `Setter`, `Destructor` types to `eviltypes`.
 - Added CHANGELOG.md file.
+- Added `evilstructs` package & `evilstructs.frozendict` class.